All toolbars and buttons moved to states/Battle/ directory
